# src/backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: REST API for AI Category Manager
  title: AI Category Manager API
  version: "1.0"
paths:
...
  /dashboard/category-breakdown:
    get:
      description: Retrieve sales breakdown by product category for the last 3 months
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category breakdown
      tags:
      - dashboard
  /dashboard/kpis:
    get:
      description: Retrieve high-level key performance indicators (GMV, Margin, Units,
        etc.)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get dashboard KPIs
      tags:
      - dashboard
  /dashboard/regional-performance:
    get:
      description: Retrieve sales performance by region for the last 3 months
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get regional performance
      tags:
      - dashboard
  /dashboard/sales-trend:
    get:
      description: Retrieve monthly sales trend data for the last 12 months
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sales trend
      tags:
      - dashboard
  /dashboard/top-products:
    get:
      description: Retrieve top 10 performing products by revenue for the last 3 months
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get top products
      tags:
      - dashboard
swagger: "2.0"